backend-deploy/requirements-replit-gemini.txt
fastapi==0.123.10
uvicorn==0.30.0
pydantic==2.9.0
google-generativeai==0.8.5